refactor(data): expose ANALYZE maintenance entrypoints via @/data/maintenance
`runAnalyzeIfStale` / `runAnalyzeNow` are welded to clientSchema bootstrap
internals (blocks count, sqlite_stat1 baseline, drift thresholds), so the
implementation stays put; add a thin public `@/data/maintenance` surface
that exposes them and repoint the roam-import / db-maintenance plugins off
the internals path.

Part of #139 (promotion-first pass; no lint rule).
